[Neat URL] Fix "=" at the end of an url which contains the char "?" before the char "!"
# Bug
If any url which uses this syntax "protocol://domain/[something]?[something]![something]" is parsed by Neat URL it will be converted into "protocol://domain/[something]?[something]![something]=" (where [something] means any string or nothing). This can be a problem for sites that don't use the syntax "?parameter=value" but have "?" and "!" in their url.
The bug occurs also with chars "(",")","?","^","$","\","@","[","]","{","}","|",";" instead of "!".
# Fix
A simple hack that deletes the last char of the new url if it is equal to "=" and different from the last one of the original url.